Added New Functions
Added Functions Get-XtremEvents and Get-XtremAlerts
Changes some formatting but not all. (more to come).
Added Some URI handling and specification to the New-XtremRequest
function. However I think I will completely redo this function as it
does not make since to add a crazy amount of logic to the function when
the URI can simply be passed to the function from its calling method.